# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gemini Image Chat API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface FileAttachment {
  /** Base64 encoded file data */
  data: string;
  mimeType: string;
  name: string;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export type ChatMessageType =
  (typeof ChatMessageType)[keyof typeof ChatMessageType];

export const ChatMessageType = {
  text: "text",
  image: "image",
} as const;

export type ChatMessageAspectRatio =
  (typeof ChatMessageAspectRatio)[keyof typeof ChatMessageAspectRatio];

export const ChatMessageAspectRatio = {
  "16:9": "16:9",
  "9:16": "9:16",
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
  type: ChatMessageType;
  imageUrl?: string;
  aspectRatio?: ChatMessageAspectRatio;
  timestamp?: string;
  /** Data URLs of attached files shown in the chat thread */
  attachmentUrls?: string[];
}

export interface ChatRequest {
  message: string;
  history?: ChatMessage[];
  /** Files attached to this message */
  attachments?: FileAttachment[];
}

export interface ChatResponse {
  message: string;
  role: string;
}

export type ImageGenerationRequestAspectRatio =
  (typeof ImageGenerationRequestAspectRatio)[keyof typeof ImageGenerationRequestAspectRatio];

export const ImageGenerationRequestAspectRatio = {
  "16:9": "16:9",
  "9:16": "9:16",
} as const;

/**
 * Optional reference image to edit or use as a base
 */
export type ImageGenerationRequestReferenceImage = {
  /** Base64 encoded image data */
  data: string;
  mimeType: string;
};

export interface ImageGenerationRequest {
  prompt: string;
  aspectRatio: ImageGenerationRequestAspectRatio;
  conversationContext?: string;
  /** Optional reference image to edit or use as a base */
  referenceImage?: ImageGenerationRequestReferenceImage;
}

export interface ImageGenerationResponse {
  /** Base64 encoded image data */
  imageData: string;
  mimeType: string;
  prompt: string;
  aspectRatio: string;
}

export interface ErrorResponse {
  error: string;
  details?: string;
}