# lib/api-zod/src/generated/types/chatMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gemini Image Chat API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessageAspectRatio } from "./chatMessageAspectRatio";
import type { ChatMessageRole } from "./chatMessageRole";
import type { ChatMessageType } from "./chatMessageType";

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
  type: ChatMessageType;
  imageUrl?: string;
  aspectRatio?: ChatMessageAspectRatio;
  timestamp?: string;
  /** Data URLs of attached files shown in the chat thread */
  attachmentUrls?: string[];
}