# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gemini Image Chat API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from "./chatMessage";
import type { FileAttachment } from "./fileAttachment";

export interface ChatRequest {
  message: string;
  history?: ChatMessage[];
  /** Files attached to this message */
  attachments?: FileAttachment[];
}