# lib/api-zod/src/generated/types/imageGenerationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gemini Image Chat API
 * OpenAPI spec version: 0.1.0
 */
import type { ImageGenerationRequestAspectRatio } from "./imageGenerationRequestAspectRatio";
import type { ImageGenerationRequestReferenceImage } from "./imageGenerationRequestReferenceImage";

export interface ImageGenerationRequest {
  prompt: string;
  aspectRatio: ImageGenerationRequestAspectRatio;
  conversationContext?: string;
  /** Optional reference image to edit or use as a base */
  referenceImage?: ImageGenerationRequestReferenceImage;
}