# lib/api-zod/src/generated/types/imageGenerationResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gemini Image Chat API
 * OpenAPI spec version: 0.1.0
 */

export interface ImageGenerationResponse {
  /** Base64 encoded image data */
  imageData: string;
  mimeType: string;
  prompt: string;
  aspectRatio: string;
}